requirements.txt
gdown
huggingface_hub==0.36.0
matplotlib
numpy<2.3
opencv-python
pandas
Pillow
psutil
py-cpuinfo
PyYAML
requests
scipy==1.13.1
seaborn
safetensors
thop
torch==2.5.1
torchvision==0.20.1
tqdm
transformers==4.57.1
